% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dincta-package.r
\docType{package}
\name{dincta}
\alias{dincta}
\title{Dincta: Data INtegration and Cell Type Annotation of Single Cell Transcriptomes}
\description{
Algorithm for single cell integration and cell type annotation.
}
\section{Usage}{
 

\enumerate{
\item ?DinctaMatrix to run Dincta on gene expression or PCA 
embeddings matrix.
\item ?DinctaClassification to run Dincta Classification on on gene expression or PCA 
embeddings matrix.
}
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/songtingstone/dincta/issues}
\item Read the manuscript
\href{https://www.biorxiv.org/content/tobeappear}{online}.
}
}

